ADD AUTENTICACION Y AISLAMIENTO DE DATOS

+++ src/models/paciente.model.ts
@@ -1,22 +1,22 @@
 import prisma from '../lib/prisma'
 import type { Prisma } from '../generated/prisma/client'
 
-export const getAll = async () => {
-  return await prisma.paciente.findMany()
+export const getAll = async (userId: number) => {
+  return await prisma.paciente.findMany({ where: { userId } })
 }
 
-export const getById = async (id: number) => {
-  return await prisma.paciente.findUnique({ where: { id } })
+export const getById = async (id: number, userId: number) => {
+  return await prisma.paciente.findUnique({ where: { id, userId } })
 }
 
 export const create = async (data: Prisma.PacienteCreateInput) => {
   return await prisma.paciente.create({ data })
 }
 
-export const update = async (id: number, data: Prisma.PacienteUpdateInput) => {
-  return await prisma.paciente.update({ where: { id }, data })
+export const update = async (id: number, userId: number, data: Omit<Prisma.PacienteUpdateInput, 'userId'>) => {
+  return await prisma.paciente.update({ where: { id,userId }, data })
 }
 
-export const remove = async (id: number) => {
-  return await prisma.paciente.delete({ where: { id } })
+export const remove = async (id: number, userId: number) => {
+  return await prisma.paciente.delete({ where: { id, userId } })
 }
